refactor: 댓글 모듈 hook 디렉토리 리팩터링
