+++ frontend/reply-module/src/__test__/intergration/adminUserComment.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, render, waitFor } from "@testing-library/react";
 import { getPasswordConfirmResult } from "../../api/getPasswordConfirmResult";
 import CommentList from "../../components/organisms/CommentList";
 import { useDeleteComment, useEditComment } from "../../hooks";
-import { useLikeComment } from "../../hooks/useLikeComment";
+import { useLikeComment } from "../../hooks/api/comment/useLikeComment";
 import { Comment } from "../../types";
 import { comments as _comments } from "../fixture/comments";
 import { myProject } from "../fixture/project";
