+++ frontend/reply-module/src/hooks/api/comment/useCreateComment.ts
@@ -1,9 +1,9 @@
 import { useMutation, useQueryClient } from "react-query";
-import { QUERY } from "../constants/api";
-import { request } from "../utils/request";
-import { Comment, CreateCommentRequestData } from "../types/comment";
-import { REACT_QUERY_KEY } from "../constants/reactQueryKey";
-import { AlertError } from "../utils/Error";
+import { QUERY } from "../../../constants/api";
+import { request } from "../../../utils/request";
+import { Comment, CreateCommentRequestData } from "../../../types/comment";
+import { REACT_QUERY_KEY } from "../../../constants/reactQueryKey";
+import { AlertError } from "../../../utils/Error";
 import axios from "axios";
 
 const _createComment = async (_data: CreateCommentRequestData) => {
