+++ frontend/reply-module/src/hooks/index.ts
@@ -1,9 +1,9 @@
-export { useUser } from "./useUser";
-export { useCreateComment } from "./useCreateComment";
-export { useEditComment } from "./useEditComment";
-export { useDeleteComment } from "./useDeleteComment";
-export { useInput } from "./useInput";
-export { useGetProject } from "./useGetProject";
-export { useLikeComment } from "./useLikeComment";
-export { useContentEditable } from "./useContentEditable";
-export { useGetAllComments } from "./useGetAllComments";
+export { useUser } from "./api/user/useUser";
+export { useCreateComment } from "./api/comment/useCreateComment";
+export { useEditComment } from "./api/comment/useEditComment";
+export { useDeleteComment } from "./api/comment/useDeleteComment";
+export { useInput } from "./ui/useInput";
+export { useGetProject } from "./api/project/useGetProject";
+export { useLikeComment } from "./api/comment/useLikeComment";
+export { useContentEditable } from "./ui/useContentEditable";
+export { useGetAllComments } from "./api/comment/useGetAllComments";
